Add the “you” badge in the sidebar
