+++ apps/token-manager/app/src/components/SideBar.js
@@ -4,6 +4,7 @@ import { IdentityBadge, Text, Viewport, breakpoint, theme } from '@aragon/ui'
 import provideNetwork from '../provide-network'
 import { formatBalance, stakesPercentages } from '../utils'
 import TokenBadge from './TokenBadge'
+import You from './You'
 
 const DISTRIBUTION_ITEMS_MAX = 7
 const DISTRIBUTION_COLORS = [
@@ -47,6 +48,7 @@ class SideBar extends React.Component {
       tokenName,
       tokenSupply,
       tokenSymbol,
+      userAccount,
       ...rest
     } = this.props
     const stakes = displayedStakes(holders, tokenSupply)
@@ -102,25 +104,22 @@ class SideBar extends React.Component {
               />
             ))}
           </StakesBar>
-          <Viewport>
-            {({ width, above }) => (
-              <ul>
-                {stakes.map(({ name, stake, color }) => (
-                  <StakesListItem key={name}>
-                    <span>
-                      <StakesListBullet style={{ background: color }} />
-                      <IdentityBadge
-                        entity={name}
-                        networkType={network.type}
-                        shorten={above('medium') || width < 520}
-                      />
-                    </span>
-                    <strong>{stake}%</strong>
-                  </StakesListItem>
-                ))}
-              </ul>
-            )}
-          </Viewport>
+          <ul>
+            {stakes.map(({ name, stake, color }) => (
+              <StakesListItem key={name}>
+                <span>
+                  <StakesListBullet style={{ background: color }} />
+                  <IdentityBadge
+                    entity={name}
+                    networkType={network.type}
+                    connectedAccount={name === userAccount}
+                  />
+                  {name === userAccount && <You />}
+                </span>
+                <strong>{stake}%</strong>
+              </StakesListItem>
+            ))}
+          </ul>
         </Part>
       </Main>
     )
